+++ CHANGELOG.md
@@ -9,6 +9,8 @@
  - Chef solo `roles_path` and `data_bags_path` can only be be single paths. [GH-446]
  - Fix `virtualbox_not_detected` error message to require 4.1.x. [GH-458]
  - Add shortname (`hostname -s`) for hostname setting on RHEL systems. [GH-456]
+ - `vagrant ssh -c` output no longer has a prefix and respects newlines
+   from the output. [GH-462]
 
 ## 0.8.2 (July 22, 2011)